- Update karma package. - Move unit test to specific folder.
