+++ package.json
@@ -46,12 +46,12 @@
     "codelyzer": "^6.0.1",
     "jasmine-core": "^3.6.0",
     "jasmine-spec-reporter": "^6.0.0",
-    "karma": "^5.2.3",
+    "karma": "^6.3.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "^4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
-    "ng-packagr": "^11.0.3",
+    "ng-packagr": "^11.2.4",
     "protractor": "^7.0.0",
     "ts-node": "^9.0.0",
     "tslint": "^6.1.3",
